Extract logic to read and validate access token to common library
