+++ src/Shark.AuthorizationServer.Core/ApplicationServices/IntrospectApplicationService.cs
@@ -15,15 +15,13 @@ using Shark.AuthorizationServer.DomainServices.Configurations;
 namespace Shark.AuthorizationServer.Core.ApplicationServices;
 
 public sealed class IntrospectApplicationService(
+    ICustomAccessTokenHandler customAccessTokenHandler,
     IRevokeTokenRepository revokeTokenRepository,
-    ICertificateValidator certificateValidator,
-    SecurityKey securityKey,
     IOptions<AuthorizationServerConfiguration> options,
     ILogger<IntrospectApplicationService> logger) : IIntrospectApplicationService
 {
+    private readonly ICustomAccessTokenHandler _customAccessTokenHandler = customAccessTokenHandler;
     private readonly IRevokeTokenRepository _revokeTokenRepository = revokeTokenRepository;
-    private readonly ICertificateValidator _certificateValidator = certificateValidator;
-    private readonly SecurityKey _securityKey = securityKey;
     private readonly AuthorizationServerConfiguration _configuration = options.Value;
     private readonly ILogger<IntrospectApplicationService> _logger = logger;
 
@@ -40,89 +38,32 @@ public sealed class IntrospectApplicationService(
             throw new InvalidOperationException("Client identifier can not be found");
         }
 
-        try
+        var jwtToken = ReadToken(request.Token);
+        if (jwtToken is null)
         {
-            var jwtToken = TryReadToken(request.Token);
-            if (jwtToken is null)
-            {
-                return new IntrospectInternalResponse { Active = false };
-            }
-
-            var response = await CheckRevokationList(jwtToken);
-            if (response != null)
-            {
-                return response;
-            }
-
-            return CreateResponse(jwtToken, clientId);
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "{Message}", ex.Message);
             return new IntrospectInternalResponse { Active = false };
         }
-    }
 
-    private JwtSecurityToken? TryReadToken(string token)
-    {
-        var handler = new JwtSecurityTokenHandler();
-        if (!handler.CanReadToken(token))
+        var response = await CheckRevokationList(jwtToken);
+        if (response != null)
         {
-            _logger.LogWarning("Token is not a well formed access token");
-            return null;
+            return response;
         }
 
-        var jwtToken = handler.ReadJwtToken(token);
-        if (!ValidateToken(handler, token))
-        {
-            _logger.LogWarning("Token is not a valid access token");
-            return null;
-        }
-
-        return jwtToken;
+        return CreateResponse(jwtToken, clientId);
     }
 
-    private bool ValidateToken(JwtSecurityTokenHandler handler, string token)
+    private JwtSecurityToken? ReadToken(string token)
     {
-        var validationParameters = new TokenValidationParameters
+        var tokenValidationParameters = new TokenValidationParameters
         {
             ValidateIssuer = true,
             ValidIssuer = _configuration.Issuer,
             ValidateAudience = false,
             ValidateIssuerSigningKey = true,
-            IssuerSigningKeyValidator = IssuerSigningKeyValidator(),
-            IssuerSigningKey = _securityKey,
-            ValidateLifetime = true,
-            ClockSkew = TimeSpan.FromSeconds(10),
         };
 
-        try
-        {
-            handler.ValidateToken(token, validationParameters, out SecurityToken validatedToken);
-            if (validatedToken is JwtSecurityToken)
-            {
-                return true;
-            }
-        }
-        catch (Exception ex)
-        {
-            _logger.LogError(ex, "{Message}", ex.Message);
-        }
-
-        return false;
-    }
-
-    private IssuerSigningKeyValidator IssuerSigningKeyValidator()
-    {
-        return (securityKey, securityToken, validationParameters) =>
-        {
-            if (securityKey is X509SecurityKey x509SecurityKey)
-            {
-                return _certificateValidator.IsValid(x509SecurityKey.Certificate);
-            }
-
-            return true;
-        };
+        return _customAccessTokenHandler.Read(token, tokenValidationParameters);
     }
 
     private async Task<IntrospectInternalResponse?> CheckRevokationList(JwtSecurityToken jwtToken)
