+++ src/Shark.AuthorizationServer.Sdk/Shark.AuthorizationServer.Sdk.csproj
@@ -11,7 +11,6 @@
   <ItemGroup>
     <PackageReference Include="Polly.Core" Version="8.4.2" />
     <PackageReference Include="Polly.Extensions" Version="8.4.2" />
-    <PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.2.0" />
     <PackageReference Include="StyleCop.Analyzers" Version="1.1.118">
       <PrivateAssets>all</PrivateAssets>
       <IncludeAssets>runtime; build; native; contentfiles; analyzers; buildtransitive</IncludeAssets>
